Address withLock review feedback
- move Agency-facing withLock API to std::concurrency
- clarify lock wait diagnostics and remove timeout dead code
- document deadlock limits and fork-inside-lock semantics
- add fork-inside-lock regression coverage
