+++ packages/agency-lang/tests/agency/lock/fork-inside-lock.test.json
@@ -1,0 +1,15 @@
+{
+  "tests": [
+    {
+      "nodeName": "main",
+      "input": "",
+      "expectedOutput": "3",
+      "evaluationCriteria": [
+        {
+          "type": "exact"
+        }
+      ],
+      "description": "fork branches can run inside withLock when they do not reacquire the same lock"
+    }
+  ]
+}
